+++ pages/dashboard/index.tsx
@@ -1,17 +1,17 @@
 import React from "react";
-import { useSession } from "next-auth/react";
-import { useEffect } from "react";
-import { useRouter } from "next/router";
+// import { useSession } from "next-auth/react";
+// import { useEffect } from "react";
+// import { useRouter } from "next/router";
 
 const Dashboard = () => {
-  const { data: session } = useSession();
-  const router = useRouter();
+  // const { data: session } = useSession();
+  // const router = useRouter();
 
-  useEffect(() => {
-    if (!session) {
-      router.push("/signin");
-    }
-  }, [session, router]);
+  // useEffect(() => {
+  //   if (!session) {
+  //     router.push("/signin");
+  //   }
+  // }, [session, router]);
   return (
     <div>
       <div className='flex justify-evenly items-center mx-10 mb-5 mt-10'>
